% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recommend.R
\name{install_sklearn}
\alias{install_sklearn}
\title{install_sklearn}
\usage{
install_sklearn(method = "auto", conda = "auto")
}
\value{
The cleaned dataset
}
\description{
Installs scikit-learn into reticluate python enviornment
}
\details{
Test out running Python
}
