% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{math_placement}
\alias{math_placement}
\title{Math Placement Exam Results}
\format{A dataset with 2696 rows and 15 columns

\describe{
   \item{Student}{Identification number for each student}
   \item{Gender}{0=Female, 1=Male}
   \item{PSATM}{PSAT score in Math}
   \item{SATM}{SAT score in Math}
   \item{ACTM}{ACTM score in Math}
   \item{Rank}{Adjusted rank in HS class}
   \item{Size}{Number of students in HS class}
   \item{GPAadj}{Adjusted GPA}
   \item{PlcmtScore}{Score on math placement exam}
   \item{Recommends}{Recommended course: R0 R01 R1 R12 R2 R3 R4 R6 R8}
   \item{Grade}{Course grade}
   \item{Rectaken}{1=recommended course, 0=otherwise}
   \item{TooHigh}{1=took course above recommended, 0=otherwise}
   \item{TooLow}{1=took course above recommended, 0=otherwise}
   \item{CourseSuccess}{1=B or better grade, 0=grade below B}
   }}
\source{
\url{http://vincentarelbundock.github.io/Rdatasets/doc/Stat2Data/MathPlacement.html}
}
\description{
Response is courseSucess
}
