% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{psc}
\alias{psc}
\title{699 Credit Card Fraud data set}
\format{A \code{tibble} with x rows and x columns

\describe{
   \item{col1}{something}
   \item{col2}{something}
   \item{col3}{something}
   }}
\description{
A very large dataset
}
