% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{recall}
\alias{recall}
\title{Recall data set}
\format{A \code{tibble} with 15 rows and 16 columns

\describe{
   \item{Name}{The name of a particular dataset}
   \item{Rows}{}
   \item{Columns}{}
   \item{Rows-Cols Ratio}{}
   \item{Number Discrete}{}
   \item{Max num factors}{}
   \item{Min num factors}{}
   \item{Avg num factors}{}
   \item{Number Continuous}{}
   \item{Gradient-Avg}{}
   \item{Gradient-Min}{}
   \item{Gradient-Max}{}
   \item{Gradient-Std}{}
   \item{SVM}{Recall when using Support Vector Machine classifier on dataset}
   \item{KNN}{Recall when using K Nearest Neighbors classifier on dataset}
   \item{NB}{Recall when using Naiive Bayes classifier on dataset}
   }}
\description{
#' A dataset containing the metafeatures snd recall of the algorithms SVM, KNN, and naiive bayes classifier of 15 datasets. These datasets are used to construct the meta model
}
