% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recommend.R
\name{recommend2}
\alias{recommend2}
\title{recommend2}
\usage{
recommend2(df, response, meta = meta_features, metric = recall)
}
\arguments{
\item{df}{Dataset in tabular form}

\item{response}{The target column of the dataset}

\item{meta_features}{The extracted meta features of various datasets}

\item{recall}{The known recall of the datasets in the meta_features dataframe}
}
\value{
A collection that contains the cleaned dataframe and the recommended algorithm as a string
}
\description{
Recommend a meta model for a classification dataset
}
\examples{
out<- recommend2(math_placement, "CourseSuccess")
View(out[[1]])
print(out[[2]])
}
