% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recommend.R
\name{recommend_alg}
\alias{recommend_alg}
\title{recommend_alg}
\usage{
recommend_alg(df, response, meta = meta_features, metric = recall)
}
\arguments{
\item{df}{The dataframe loaded  by the user}

\item{response}{The target column of thedataframe loaded  by the user}

\item{meta_features}{The preextracted meta features of various datasets}

\item{recall}{The known recall of the datasets in the meta_features dataframe}
}
\value{
The reccomended algorithm as a string
}
\description{
Recommend a meta model for a classification dataset. Used internally by R shiny app
}
