% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recommend.R
\name{recommend_shiny}
\alias{recommend_shiny}
\title{recommend}
\usage{
recommend_shiny(df, response, meta = meta_features, metric = recall)
}
\arguments{
\item{sampsize}{The size of the sample}

\item{nsim}{The number of simulations}
}
\value{
A seaborn kdeplot (kernel density estimate plot)
}
\description{
Recommend a meta model for a classification dataset
}
\details{
Test out running Python
}
