% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{urine}
\alias{urine}
\title{Urine data set}
\format{A \code{tibble} with 79 rows and 7 columns

\describe{
   \item{r}{Indicator of the presence of calcium oxalate crystals.}
   \item{gravity}{The specific gravity of the urine}
   \item{ph}{The pH reading of the urine.}
   \item{osmo}{The osmolarity of the urine. Osmolarity is proportional to the concentration of molecules in solution}
   \item{cond}{The conductivity of the urine. Conductivity is proportional to the concentration of charged ions in solution.}
   \item{urea}{The urea concentration in millimoles per litre.}
   \item{calc}{The calcium concentration in millimoles per litre.}
   }}
\source{
\url{http://vincentarelbundock.github.io/Rdatasets/doc/boot/urine.html}
}
\description{
A very large dataset
}
