% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recommend.R
\name{recommend}
\alias{recommend}
\title{recommend}
\usage{
recommend(meta = meta_features, metric = recall)
}
\arguments{
\item{meta_features}{The extracted meta features of various datasets}

\item{recall}{The known recall of the datasets in the meta_features dataframe}
}
\value{
A collection that contains the cleaned dataframe and the recommended algorithm as a string
}
\description{
Recommend a meta model for a classification dataset
}
\examples{
out<- recommend()
View(out[[1]])
print(out[[2]])
}
